Added detail view and restaurant features by id
